Add canonical hierarchical routes for desktop windows
